@@ -2,5 +2,9 @@ package br.com.bdws.start_spring.config.security;
 
 public enum Authorities {
     EDITOR,
-    ADMIN
+    ADMIN;
+
+    public  String role() {
+        return "ROLE_".concat(name());
+    }
 }
