@@ -1,17 +1,17 @@
 package br.com.bdws.start_spring.config.security;
 
+import jakarta.servlet.FilterChain;
+import jakarta.servlet.ServletException;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
-import org.springframework.web.filter.GenericFilterBean;
+import org.springframework.web.filter.OncePerRequestFilter;
 
-import javax.servlet.FilterChain;
-import javax.servlet.ServletException;
-import javax.servlet.ServletRequest;
-import javax.servlet.ServletResponse;
-import javax.servlet.http.HttpServletRequest;
 import java.io.IOException;
+import java.util.Objects;
 
-public class JWTAuthenticationFilter extends GenericFilterBean {
+public class JWTAuthenticationFilter  extends OncePerRequestFilter {
 
     private TokenAuthenticationService tokenAuthenticationService;
 
@@ -20,13 +20,11 @@ public class JWTAuthenticationFilter extends GenericFilterBean {
     }
 
     @Override
-    public void doFilter(ServletRequest request, ServletResponse response, FilterChain filterChain)
-            throws IOException, ServletException {
-
-        Authentication authentication = tokenAuthenticationService
-                .getAuthentication((HttpServletRequest) request);
-
-        SecurityContextHolder.getContext().setAuthentication(authentication);
+    protected void doFilterInternal(HttpServletRequest request, HttpServletResponse response, FilterChain filterChain) throws ServletException, IOException {
+        Authentication authentication = tokenAuthenticationService.getAuthentication(request);
+        if (Objects.nonNull(authentication)) {
+            SecurityContextHolder.getContext().setAuthentication(authentication);
+        }
         filterChain.doFilter(request, response);
     }
 }
